+++ app/apis/catalog/hca-atlas-tracker/aws/schemas.ts
@@ -1,0 +1,59 @@
+import { array, InferType, number, object, string } from "yup";
+
+// AWS S3 and SNS Event Validation Schemas
+// These schemas validate the structure of AWS events received via SNS notifications
+
+export const s3ObjectSchema = object({
+  eTag: string().required(),
+  key: string().required(),
+  size: number().required(),
+  userMetadata: object()
+    .shape({
+      "source-sha256": string()
+        .matches(
+          /^[a-fA-F0-9]{64}$/,
+          "SHA256 must be a 64-character hexadecimal string"
+        )
+        .required("SHA256 metadata is required for file integrity validation"),
+    })
+    .required(),
+  versionId: string().nullable().optional(),
+}).required();
+
+export const s3BucketSchema = object({
+  name: string().required(),
+}).required();
+
+export const s3RecordSchema = object({
+  eventName: string().required(),
+  eventSource: string().oneOf(["aws:s3"]).required(),
+  eventTime: string().required(),
+  s3: object({
+    bucket: s3BucketSchema,
+    object: s3ObjectSchema,
+  }).required(),
+}).required();
+
+export const s3EventSchema = object({
+  Records: array().of(s3RecordSchema).min(1).required(),
+}).required();
+
+export const snsMessageSchema = object({
+  Message: string().required(),
+  MessageId: string().required(),
+  Signature: string().required(),
+  SignatureVersion: string().required(),
+  SigningCertURL: string().url().required(),
+  Subject: string().optional(),
+  SubscribeURL: string().url().optional(),
+  Timestamp: string().required(),
+  Token: string().optional(),
+  TopicArn: string().required(),
+  Type: string().oneOf(["Notification"]).required(),
+  UnsubscribeURL: string().url().optional(),
+}).required();
+
+// Type inference from Yup schemas
+export type S3EventRecord = InferType<typeof s3RecordSchema>;
+export type S3Event = InferType<typeof s3EventSchema>;
+export type SNSMessage = InferType<typeof snsMessageSchema>;
